improve: restructure directory
