+++ src/views/RewardsProgram/GenericRewardsProgram/GenericCard.tsx
