+++ src/views/RewardsProgram/GenericRewardsProgram/GenericInformationCard.tsx
